KIEV21DEVS-141 captcha

+++ backend/src/main/java/com/eshop/backend/auth/dao/email/EmailTokenDaoImpl.java
@@ -55,7 +55,7 @@ public class EmailTokenDaoImpl implements EmailTokenDao {
     }
 
     @Override
-    public EmailTokenModel getById(int id) {
+    public EmailTokenModel getById(Long id) {
         return null;
     }
